Remove readdir cookies and set dirent.d_off in NFS
git-svn-id: https://socsvn.freebsd.org/socsvn/soc2011/gk/ino64-head@224050 fcd59f9a-a76a-e011-b4ea-001e4f0fb9b1

+++ sys/fs/nfs/nfs.h
@@ -251,7 +251,7 @@ struct nfsreferral {
 	u_char		*nfr_srvlist;	/* List of servers */
 	int		nfr_srvcnt;	/* number of servers */
 	vnode_t		nfr_vp;	/* vnode for referral */
-	u_int32_t	nfr_dfileno;	/* assigned dir inode# */
+	u_int64_t	nfr_dfileno;	/* assigned dir inode# */
 };
 
 /*
